clh: Deal with devices that require IOMMU
Let's take advantage of the already present `IOMMU` property for the
hypervisors and, using that, let's do the right thing on Cloud
Hypervisor when that's used.

Mind that for now I'm *not* exposing that as a configuration option as
we do *not* want users blindly using that.  However, if someone happens
to have a special device that must be plugged behind a different PCI
segment using IOMMU, let's allow the user to do that via the already
exposed annotations.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>